requirements.txt
colorama
fpdf2
pygments
pdfkit